[abi & utils]: new crate abi and moved stuff from utils to abi
