+++ abi/README.md
@@ -1,0 +1,4 @@
+# safa-abi
+an ABI over the [SafaOS's](https://github.com/SafaOS/SafaOS) kernel syscalls (numbers only) and errors.
+
+you must likely want to use the [SafaOS's api](https://github.com/SafaOS/safa-api) instead.
