+++ utils/Cargo.toml
@@ -1,19 +1,8 @@
 [package]
 name = "safa-utils"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
-arrayvec = { version = "0.7.6", git = "https://github.com/ObserverUnit/arrayvec", default-features = false }
-
-core = { version = "1.0.0", optional = true, package = "rustc-std-workspace-core" }
-alloc = { version = "1.0.0", optional = true, package = "rustc-std-workspace-alloc" }
-compiler_builtins = { version = "0.1", optional = true }
-[features]
-default = []
-rustc-dep-of-std = [
-    "core",
-    "alloc",
-    "compiler_builtins/rustc-dep-of-std",
-    "arrayvec/rustc-dep-of-std",
-]
+arrayvec = { version = "0.7.6", default-features = false }
+safa-abi = { path = "../abi" }
